[XLA:GPU] Fix crash in reduction emitter. Use `int64_t` to avoid an overflows.
`CalculateVirtualThreadScalingFactorForReduction` should trigger for small row reductions only. An integer overflow caused it to trigger for very large row reductions and caused runtime crashes.

See https://github.com/google/jax/issues/14013

PiperOrigin-RevId: 511150718

+++ tensorflow/compiler/xla/service/gpu/ir_emitter_unnested.cc
@@ -4628,7 +4628,7 @@ static int GetPrimitiveBitwidth(mlir::Value i) {
 static int CalculateVirtualThreadScalingFactorForReduction(
     const ReductionDimensions& reduction_dimensions,
     const se::CudaComputeCapability& cc) {
-  int dimx = reduction_dimensions.dimensions[kDimX];
+  int64_t dimx = reduction_dimensions.dimensions[kDimX];
   if (reduction_dimensions.is_row_reduction && dimx <= 128) {
     int rows_per_warp = RowReductionGetRowsPerWarp(dimx);
     if (cc.IsAtLeast(se::CudaComputeCapability::AMPERE)) {
